*** empty log message ***

+++ src/org/yccheok/jstock/gui/PortfolioManagementJPanel.java
@@ -1607,9 +1607,11 @@ public class PortfolioManagementJPanel extends javax.swing.JPanel {
                     depositSummaryFile.delete();
                     dividendSummaryFile.delete();
                     final File config_directory = new File(org.yccheok.jstock.gui.Utils.getUserDataDirectory() + country + File.separator + "config");
-                    if (config_directory.list().length <= 0) {
-                        // Empty directory. Remove it.
-                        config_directory.delete();
+                    if (config_directory.isDirectory()) {
+                        if (config_directory.list().length <= 0) {
+                            // Empty directory. Remove it.
+                            config_directory.delete();
+                        }
                     }
                 }
             }
